# apps/frontend/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: StipendieAssistenten Backend API
  description: |
    Required backend endpoints for StipendieAssistenten frontend.
    All authenticated endpoints require a Bearer token in the Authorization header.
  version: 1.0.0
servers:
  - url: /api
    description: Default API path (configure via VITE_API_URL env variable)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
          description: Unique user identifier
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
        avatar:
          type: string
          format: uri
          nullable: true
    AuthTokens:
      type: object
      required:
        - access_token
        - user
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Optional refresh token
        user:
          $ref: '#/components/schemas/User'
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
paths:
...
  /auth/login:
    post:
      summary: Login with email and password
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/signup:
    post:
      summary: Register a new user
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                  minLength: 8
                name:
                  type: string
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '400':
          description: Invalid input or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      summary: Get current authenticated user
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated or token expired
  /auth/logout:
    post:
      summary: Logout current user (optional server-side session invalidation)
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successfully logged out
        '401':
          description: Not authenticated
...
  /grants:
    get:
      summary: List available grants
      description: Public endpoint - no authentication required
      tags: [Grants]
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of grants
          content:
            application/json:
              schema:
                type: object
                properties:
                  grants:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        organization:
                          type: string
                        summary:
                          type: string
                        category:
                          type: string
                        deadline:
                          type: string
                          format: date
                          nullable: true
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /grants/{id}:
    get:
      summary: Get grant details
      description: Public endpoint - no authentication required
      tags: [Grants]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Grant details
        '404':
          description: Grant not found

  # ============================================
  # SAVED/FAVORITE GRANTS (Authenticated)
  # ============================================
...
  /applications:
    get:
      summary: Get user's grant applications
      tags: [Applications]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    grant_id:
                      type: string
                    grant_name:
                      type: string
                    status:
                      type: string
                      enum: [draft, submitted, approved, rejected]
                    created_at:
                      type: string
                      format: date-time
                    updated_at:
                      type: string
                      format: date-time
        '401':
          description: Not authenticated

    post:
      summary: Create a new application
      tags: [Applications]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - grant_id
              properties:
                grant_id:
                  type: string
                content:
                  type: string
      responses:
        '201':
          description: Application created
        '401':
          description: Not authenticated
  /applications/{id}:
    get:
      summary: Get application details
      tags: [Applications]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Application details
        '401':
          description: Not authenticated
        '404':
          description: Application not found

    put:
      summary: Update application
      tags: [Applications]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                status:
                  type: string
                  enum: [draft, submitted]
      responses:
        '200':
          description: Application updated
        '401':
          description: Not authenticated

    delete:
      summary: Delete application
      tags: [Applications]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Application deleted
        '401':
          description: Not authenticated

  # ============================================
  # AI GENERATION (Premium feature)
  # ============================================
  /generate/application:
    post:
      summary: Generate AI-assisted application text (Premium feature)
      description: |
        Requires premium subscription or available credits.
        Uses family profile data to personalize the generated application.
      tags: [Generation, Premium]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - grant_id
              properties:
                grant_id:
                  type: string
                additional_context:
                  type: string
                  description: Optional additional information from user
      responses:
        '200':
          description: Generated application text
          content:
            application/json:
              schema:
                type: object
                properties:
                  generated_text:
                    type: string
                  credits_remaining:
                    type: integer
                    nullable: true
        '401':
          description: Not authenticated
        '402':
          description: Premium subscription required or no credits remaining
        '404':
          description: Grant not found